Authentication Work
Started adding firebase authentication.
Running up against this error message in the console when I tried to test it however. Need to troubleshoot.

"Either AngularFireModule has not been provided in your AppModule (this can be done manually or implictly using
provideFirebaseApp) or you're calling an AngularFire method outside of an NgModule (which is not supported)."